Remove useResizeObserver util and refactor into useCallback hook
Refactored useResizeObserver into more compact and effective useCallback
hook in order to fix bug with adding the 'unmountOnExit' prop to MUI
Collapse components (adding 'unmountOnExit' prevented the treemap from
re-mounting when the Collapse was opened again). Now the collapse components for the treemaps will
mount and unmount, which helps fix the animation speed of the Drawer
while the treemaps are collapsed.